+++ stxt/out/core/Parser.js
@@ -99,7 +99,10 @@ class Parser {
             // Pasamos validators
             this.validators.forEach(validator => {
                 try {
-                    validator.validate(completed);
+                    const errors = validator.validate(completed);
+                    errors.forEach(error => {
+                        result.addError(error);
+                    });
                 }
                 catch (e) {
                     this.handleError(e, completed.getLine(), result, "VALIDATION_ERROR", "UNKNOWN_VALIDATION_ERROR");