+++ stxt/src/processors/Validator.ts
@@ -1,5 +1,6 @@
 import { Node } from "../core/Node";
+import { ValidationException } from "../exceptions/ValidationException";
 
 export interface Validator {
-	validate(n: Node): void;
+	validate(n: Node): ValidationException[];
 }